requirements.txt
flask
slack_sdk
openai
python-dotenv
